task-tracker-lab1/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic[email]
SQLAlchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv==1.0.0
redis
bcrypt
